Use BaseConnectorTests in PostgreSql connector
